# doc/openapi.yaml
openapi: 3.0.0
info:
  title: Battery Analysis Platform API
  description: |
    储能电池寿命分析及算法测试平台 API 规范

    核心功能:
    - 用户认证与授权 (JWT)
    - 数据管理与可视化分析
    - 训练平台 (Baseline/BiLSTM/DeepHPM)
    - 测试平台 (模型推理评估)
    - 模型版本管理
    - WebSocket实时进度推送
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: 本地开发服务器
...
paths:
...
  /v1/testing/jobs:
    post:
      tags: [Testing]
      summary: 创建测试任务
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                model_version_id:
                  type: integer
                dataset_id:
                  type: integer
                target:
                  type: string
                battery_ids:
                  type: array
                  items:
                    type: integer
                horizon:
                  type: integer
      responses:
        "201":
          description: 创建成功

    get:
      tags: [Testing]
      summary: 列出测试任务
      security:
        - BearerAuth: []
      parameters:
        - name: status_filter
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: 成功
  /v1/testing/jobs/{job_id}:
    get:
      tags: [Testing]
      summary: 获取测试任务详情
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
...
  /v1/testing/jobs/{job_id}/predictions:
    get:
      tags: [Testing]
      summary: 获取预测结果
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
        - name: battery_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: 成功
  /v1/testing/jobs/{job_id}/logs:
    get:
      tags: [Testing]
      summary: 获取测试日志（从文件读取）
      description: |
        从日志文件读取测试日志

        返回格式:
        - job_id: 测试任务ID
        - log_file_path: 日志文件路径
        - total_lines: 文件总行数
        - returned_lines: 返回的日志行数
        - logs: 日志数组，每条包含 timestamp, level, message

        注意: 日志现在存储在文件中，不再存储在数据库
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
        - name: level
          in: query
          description: 按日志级别过滤 (DEBUG/INFO/WARNING/ERROR)
          schema:
            type: string
        - name: limit
          in: query
          description: 返回最新的N条日志，默认1000
          schema:
            type: integer
            default: 1000
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: integer
                  log_file_path:
                    type: string
                    description: 日志文件相对路径
                  total_lines:
                    type: integer
                    description: 文件总行数
                  returned_lines:
                    type: integer
                    description: 返回的日志行数
                  logs:
                    type: array
                    items:
                      type: object
                      properties:
                        timestamp:
                          type: string
                        level:
                          type: string
                        message:
                          type: string
                  message:
                    type: string
                    description: 当日志文件不存在时的提示信息
  /v1/testing/jobs/{job_id}/export:
    post:
      tags: [Testing]
      summary: 导出测试结果
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
        - name: format
          in: query
          schema:
            type: string
            enum: [CSV, XLSX]
      responses:
        "200":
          description: 导出任务已创建

  # ==========================================
  # WebSocket
  # ==========================================
...
  /v1/testing/ws/jobs/{job_id}:
    get:
      tags: [WebSocket]
      summary: 测试任务WebSocket
      description: |
        实时接收测试进度

        消息类型:
        - log: 日志消息（仅推送 WARNING 和 ERROR 级别）
        - progress: 阶段进度
        - status_change: 状态变化

        注意:
        - 完整日志存储在文件中，可通过日志API查询
        - WebSocket仅推送关键日志（WARNING/ERROR）以减少网络流量

        连接: ws://localhost:8000/api/v1/testing/ws/jobs/{job_id}
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "101":
          description: 切换到WebSocket

  # ==========================================
  # Health Check
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...